:sparkles: feat: support decimal wallet and transfer amounts
Parse and format wallet balances and transfers as fractional coin
values (double) so users can send and display decimal amounts.

+++ lib/data/models/wallet.dart
@@ -1,5 +1,5 @@
 class Wallet {
-  final int balance;
+  final double balance;
   final List<WalletTransaction> transactions;
 
   const Wallet({required this.balance, this.transactions = const []});
@@ -7,7 +7,7 @@ class Wallet {
   factory Wallet.fromJson(Map<String, dynamic> json) {
     final list = json['transactions'];
     return Wallet(
-      balance: (json['balance'] as num?)?.toInt() ?? 0,
+      balance: (json['balance'] as num?)?.toDouble() ?? 0,
       transactions: list is List
           ? list
               .whereType<Map<String, dynamic>>()
@@ -20,8 +20,8 @@ class Wallet {
 
 class WalletTransaction {
   final int id;
-  final int amount;
-  final int balanceAfter;
+  final double amount;
+  final double balanceAfter;
   final String type;
   final String description;
   final DateTime createdAt;
@@ -38,8 +38,8 @@ class WalletTransaction {
   factory WalletTransaction.fromJson(Map<String, dynamic> json) {
     return WalletTransaction(
       id: (json['id'] as num?)?.toInt() ?? 0,
-      amount: (json['amount'] as num?)?.toInt() ?? 0,
-      balanceAfter: (json['balance_after'] as num?)?.toInt() ?? 0,
+      amount: (json['amount'] as num?)?.toDouble() ?? 0,
+      balanceAfter: (json['balance_after'] as num?)?.toDouble() ?? 0,
       type: json['type'] as String? ?? '',
       description: json['description'] as String? ?? '',
       createdAt: DateTime.tryParse(json['created_at'] as String? ?? '')?.toLocal() ??
